Add logic to retain child task exit status if so configured
git-svn-id: http://svn.code.sf.net/p/nuttx/code/trunk@5553 42af7a65-404d-4744-a932-0658087f49c3

+++ nuttx/include/sys/types.h
@@ -158,13 +158,13 @@ typedef uint16_t     ino_t;
  * negative PID values are used to represent invalid PIDs.
  */
 
-typedef int          pid_t;
+typedef int16_t      pid_t;
 
 /* id_t is a general identifier that can be used to contain at least a pid_t,
  * uid_t, or gid_t.
  */
 
-typedef unsigned int id_t;
+typedef int16_t      id_t;
 
 /* Signed integral type of the result of subtracting two pointers */
